refactor(picker): handle picker errors
Slim down picker.pick function to only return the exit-status of the
picker subprocess and the index of selected. If a selection was not made
return the exception as the index allowing the caller to handle the
exception.

Refactor edit command to print error message and exit when the picker
returns with non-zero exit status or when a selection was not made.