feat(watcher): add auto-export watcher daemon
Add daemonized watcher that will be ensured running whenever any of the
commands are run. The daemon acts like a server that listens for new
connections and their messages which are paths that clients are
requesting to be watched. `create` and `edit` automatically start
watching the directory of the target figure. `watch` allows
specification of a directory to watch

Watched paths are recursive and a minimal set of watched directories is
maintained. This means that when a new path is requested to be watched,
it will be compared against currently watched directories. Overlapping
watched directories are eliminated. For example, if the requested
directory contains a directory already being watched, the directory
already being watched will be removed and the requested directory will
be watched. The other way also works--if the requested path is already
watched via one of its parents, it will not be added.