refactor(create): add relative-from option and print export path
Change the `create` command's figure-dir argument to an option, and
add the relative-from option.

Make the inclusion text be the relative path of the exported figure. The
relative-from option selects where the relative path is from and is the
CWD by default. The figure-dir is where the figure is created and is CWD
by default.